triage: address MysticatBot review on owned-urls endpoint | LLMO-6086
- bound per-project fan-out in getOwnedUrls via shared mapWithConcurrency util (cap 8) — avoids unbounded parallel Semrush calls for many-market brands

- apply category filter to the trend payload too, so weekly sparklines and aggregate totals share the same filter set

- allowlist-validate referralSource (unknown → dropped, RPC defaults to optel) instead of silently returning empty

- cap the date-range span at 366 days (defense-in-depth alongside the traffic RPC p_urls cap)
